Refactor menu ID handling to use int instead of long; update related components and methods for consistency in data types; enhance dropdown action assignment logic in inspector.

+++ Engine/TemplateGame/Component/Menu.cs
@@ -9,7 +9,7 @@ namespace TemplateGame.Component
     /// </summary>
     public class Menu : IPermanentRenderingObject
     {
-        internal long ID { get; set; }
+        internal int ID { get; set; }
         /// <summary>
         /// The absolute position on the X axis.
         /// </summary>
@@ -45,7 +45,7 @@ namespace TemplateGame.Component
         internal bool IsVisible { get; set; }
         internal Game Game { get; set; }
 
-        public Menu(Game game, long id, int xPos, int yPos, int width, int height, bool isFullScreen, List<Block> blockList, Color windowColor, Color windowBorderColor)
+        public Menu(Game game, int id, int xPos, int yPos, int width, int height, bool isFullScreen, List<Block> blockList, Color windowColor, Color windowBorderColor)
         {
             Game = game;
             ID = id;
